+++ Items/SCP178/SCP1783DVision.cs
@@ -1,4 +1,4 @@
-﻿/*using BepInEx.Logging;
+﻿using BepInEx.Logging;
 using GameNetcodeStuff;
 using System;
 using System.Collections.Generic;
@@ -148,4 +148,3 @@ namespace HeavyItemSCPs.Items.SCP178
         } 
     }
 }
-*/