+++ ThunderStore/CHANGELOG.md
@@ -1,2 +1,9 @@
+## 1.0.1
+- SCP-427-1 now only takes damage from mines if chasing a player
+- Fixed issue with players not transforming
+- Fixed some issues with the necklace
+- Player will no longer be critically injured after thrown and will lose half their stamina instead
+- Fixed death animation
+
 ## 1.0.0
 - Initial release adding SCP-427